feat: create shared logger
