+++ packages/shared-utils/package.json
@@ -8,7 +8,8 @@
   },
   "dependencies": {
     "raw-body": "2.4.1",
-    "slugify": "^1.6.6"
+    "slugify": "^1.6.6",
+    "pino": "^9.9.0"
   },
   "devDependencies": {
     "next": "^15.3.4"
